requirements.txt
deepspeech
discord.py
ffmpeg-python
halo
numpy
scipy
sounddevice
streamlink
webrtcvad
